small cleanup refactor
