+++ apps/cli/src/utils.ts
@@ -1,0 +1,10 @@
+export // TODO: just use the native one when support for node 12 gets dropped
+const allSettled: typeof Promise.allSettled = (promises: Promise<any>[]) =>
+  Promise.all(
+    promises.map((promise) =>
+      promise.then(
+        (value) => ({ status: 'fulfilled' as const, value }),
+        (reason) => ({ status: 'rejected' as const, reason }),
+      ),
+    ),
+  );
